+++ src/components/SelectBox.js
@@ -1,5 +1,6 @@
 import React, {useState} from 'react'
 import Select from 'react-select';
+import styled from 'styled-components';
 
 const options = [
     {value : '종합', label : '종합'},
@@ -17,15 +18,26 @@ const SelectBox = ({onChange, className}) => {
     }
   }
   return (
-    <div className = {className}>
+    <StyledSelectBox className = {className}>
       <Select id="selectBox"
         value={selectedOption}
         onChange={handleChange}
         options={options}
       />
 
-    </div>
+    </StyledSelectBox>
   )
 }
 
 export default SelectBox;
+
+const StyledSelectBox = styled.div`
+  padding-top : 7px;
+  width: 200px;
+  padding: 8px;
+  border-radius: 12px;
+  background-color: #ffffff;
+  align-self: center;
+  box-shadow: 0px 4px 4px rgba(0, 0, 0, 0.25);
+  cursor: pointer;
+`;
