--- src/pages/App.css
@@ -1,10 +1,0 @@
-.center {
-    text-align : center ;
-}
-#icon {
-    border-bottom : 1px solid rgb(47, 47, 47);
-    padding-bottom : 5px;
-}
-#selectBox {
-    padding-top : 7px;
-}
