+++ src/pages/Explore.js
@@ -1,0 +1,14 @@
+import React from 'react';
+import Icon from '../components/Icon';
+import Section from '../components/Section';
+
+const Explore = () => {
+  return (
+    <>
+      <Icon/>
+      <Section/>
+    </>
+  );
+}
+
+export default Explore;
