--- src/pages/Icon.js
@@ -1,10 +1,0 @@
-import React from 'react'
-import './App.css';
-
-const Icon = () => {
-  return (
-    <div id="icon" className="center">뮤미지</div>
-  )
-}
-
-export default Icon
