+++ src/pages/Landing.js
@@ -1,0 +1,14 @@
+import React from 'react'
+import styled from 'styled-components'
+
+const Landing = () => {
+  return (
+    <>
+      <StyledTitle>Landing</StyledTitle>
+    </>
+  )
+}
+
+export default Landing
+
+const StyledTitle = styled.div``;
