--- src/pages/StyledSelectBox.js
@@ -1,14 +1,0 @@
-import styled from 'styled-components';
-import SelectBox from './SelectBox'
-
-const StyledSelectBox = styled(SelectBox)`
-width: 200px;
-padding: 8px;
-border-radius: 12px;
-background-color: #ffffff;
-align-self: center;
-box-shadow: 0px 4px 4px rgba(0, 0, 0, 0.25);
-cursor: pointer;
-`;
-
-export default StyledSelectBox;
